web - fix integration tests
