+++ src/vs/platform/log/common/log.ts
@@ -595,4 +595,3 @@ export function LogLevelToString(logLevel: LogLevel): string {
 		case LogLevel.Off: return 'off';
 	}
 }
-
