+++ src/vs/workbench/services/log/electron-sandbox/logService.ts
@@ -3,13 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { LogService, ConsoleLogger, MultiplexLogService, ILogger, LogLevel, ILogService } from 'vs/platform/log/common/log';
+import { LogService, ConsoleLogger, MultiplexLogService, ILogger, LogLevel } from 'vs/platform/log/common/log';
 import { INativeWorkbenchEnvironmentService } from 'vs/workbench/services/environment/electron-sandbox/environmentService';
 import { LogLevelChannelClient, FollowerLogService, LoggerChannelClient } from 'vs/platform/log/common/logIpc';
 import { DisposableStore } from 'vs/base/common/lifecycle';
-import { CommandsRegistry } from 'vs/platform/commands/common/commands';
-import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
-import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 
 export class NativeLogService extends LogService {
 
@@ -38,20 +35,3 @@ export class NativeLogService extends LogService {
 		this._register(disposables);
 	}
 }
-
-// --- Internal commands to improve extension test runs
-
-CommandsRegistry.registerCommand('_extensionTests.setLogLevel', function (accessor: ServicesAccessor, level: number) {
-	const logService = accessor.get(ILogService);
-	const environmentService = accessor.get(IEnvironmentService);
-
-	if (environmentService.isExtensionDevelopment && !!environmentService.extensionTestsLocationURI) {
-		logService.setLevel(level);
-	}
-});
-
-CommandsRegistry.registerCommand('_extensionTests.getLogLevel', function (accessor: ServicesAccessor) {
-	const logService = accessor.get(ILogService);
-
-	return logService.getLevel();
-});
